code/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "loraforge"
version = "0.1.0"
description = "End-to-end inference optimization pipeline for LoRA fine-tuned LLMs"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Deepank Dixit", email = "deepankdixit0804@gmail.com" }]
keywords = ["llm", "inference", "optimization", "lora", "quantization", "speculative-decoding", "kv-cache"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.10"
dependencies = [
    "torch>=2.3.0",
    "transformers>=4.44.0",
    "peft>=0.12.0",
    "vllm>=0.6.0",
    "typer[all]>=0.12.0",
    "pydantic>=2.7.0",
    "pyyaml>=6.0.1",
    "streamlit>=1.38.0",
    "plotly>=5.23.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "tqdm>=4.66.0",
    "datasets>=2.21.0",
    "accelerate>=0.33.0",
    "lm-eval>=0.4.3",
    "pynvml>=11.5.0",
    "aiohttp>=3.10.0",
    "rich>=13.7.0",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
modelopt = ["nvidia-modelopt[torch]>=0.11.0"]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "black>=24.8.0",
    "isort>=5.13.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "pre-commit>=3.8.0",
]

[project.scripts]
loraforge = "activity5_benchmark_dashboard.cli.main:app"

[project.urls]
Homepage = "https://github.com/deepank-dixit/loraforge"
Repository = "https://github.com/deepank-dixit/loraforge"
Issues = "https://github.com/deepank-dixit/loraforge/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["shared*", "activity*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests requiring GPU or long runtime",
    "gpu: marks tests requiring CUDA GPU",
]
